[SofaGeneralAnimationLoop, Sofa.SimulationCore] Extract MechanicalAccumulateJacobian in its own files
